+++ instances/downloader_server.cpp
@@ -1,23 +1,23 @@
 #include <rsi/uploader/Downloader.hpp>
-#include <rsi/sockets/StatusLine.hpp>
+#include <extras/status/StatusLine.hpp>
 #include <iostream>
 
 using namespace  extras;
 
 int main(int argc, char const* argv[]) {
     try {
-        std::cout << rsi::start(argv[0]) << std::endl;
+        std::cout << extras::start(argv[0]) << std::endl;
         extras::rsi::DownloaderServer downloader;
         downloader.parameters(argc, argv);
         downloader.connect();
         downloader.transfer();
-        std::cout << rsi::pass("File data downloaded successfully") << std::endl;
+        std::cout << extras::pass("File data downloaded successfully") << std::endl;
         downloader.close();
-        std::cout << rsi::end("Closed the connection") << std::endl << std::endl;
+        std::cout << extras::end("Closed the connection") << std::endl << std::endl;
         exit(0);
     }
     catch (std::exception& ex) {
-        std::cout << rsi::fail(ex.what()) << std::endl << std::endl;
+        std::cout << extras::fail(ex.what()) << std::endl << std::endl;
         exit(-1);
     }
 }