+++ src/subsystem/configure_serversocket.cpp
@@ -2,7 +2,7 @@
 #include <unistd.h>
 #include <rsi/subsystem.hpp>
 #include <rsi/exceptions.hpp>
-#include <rsi/sockets/StatusLine.hpp>
+#include <extras/status/StatusLine.hpp>
 
 //
 // https://cdmana.com/2021/02/20210220084428841j.html
@@ -16,7 +16,7 @@ int extras::rsi::configure_serversocket(const char* ip, int port,
 
     sockfd = socket(AF_INET, SOCK_STREAM, 0);
     SocketException::assertionLTZero(sockfd, "Error in socket", __INFO__);
-    std::cout << rsi::pass("Server socket created successfully") << std::endl;
+    std::cout << extras::pass("Server socket created successfully") << std::endl;
 
     if (timeoutmode) {
         // Set a 7 second timeout
@@ -24,7 +24,7 @@ int extras::rsi::configure_serversocket(const char* ip, int port,
         timeout.tv_sec = 20;  // after 7 seconds connect() will timeout
         timeout.tv_usec = 0;
         setsockopt(sockfd, SOL_SOCKET, SO_RCVTIMEO, &timeout, sizeof(timeout));
-        std::cout << rsi::pass("Server socket times out in 20 seconds") << std::endl;
+        std::cout << extras::pass("Server socket times out in 20 seconds") << std::endl;
     }
 
     // Forcefully attaching socket to the port 8080
@@ -38,9 +38,9 @@ int extras::rsi::configure_serversocket(const char* ip, int port,
 
     int e = bind(sockfd, (struct sockaddr*)&server_addr, sizeof(server_addr));
     SocketException::assertionLTZero(e, "Error in bind", __INFO__);
-    std::cout << rsi::pass("Binding successfull") << std::endl;
+    std::cout << extras::pass("Binding successfull") << std::endl;
 
     SocketException::assertionNEZero(listen(sockfd, 10), "Error in listening", __INFO__);
-    std::cout << rsi::pass("Listening") << std::endl;
+    std::cout << extras::pass("Listening") << std::endl;
     return sockfd;
 }