+++ src/subsystem/connect_to_server.cpp
@@ -2,7 +2,7 @@
 #include <unistd.h>
 #include <rsi/subsystem.hpp>
 #include <rsi/exceptions.hpp>
-#include <rsi/sockets/StatusLine.hpp>
+#include <extras/status/StatusLine.hpp>
 
 int extras::rsi::connect_to_server(const char* ip, int port,
     struct sockaddr_in& server_addr) {
@@ -12,7 +12,7 @@ int extras::rsi::connect_to_server(const char* ip, int port,
 
     sockfd = socket(AF_INET, SOCK_STREAM, 0);
     SocketException::assertionLTZero(sockfd, "Error in socket", __INFO__);
-    std::cout << rsi::pass("Server socket created successfully") << std::endl;
+    std::cout << extras::pass("Server socket created successfully") << std::endl;
 
     server_addr.sin_family = AF_INET;
     server_addr.sin_port = port;
@@ -20,7 +20,7 @@ int extras::rsi::connect_to_server(const char* ip, int port,
 
     e = connect(sockfd, (struct sockaddr*)&server_addr, sizeof(server_addr));
     SocketException::assertionEQMinusOne(e, "Error in socket", __INFO__);
-    std::cout << rsi::pass("Connected to Server") << std::endl;
+    std::cout << extras::pass("Connected to Server") << std::endl;
 
     return sockfd;
 }