+++ changelog.md
@@ -4,8 +4,8 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).</br>
 > To add a tag to the git repository do:
-> > git tag -a v2.15.0 -m "Added ColorBar"</br>
-> > git push origin v2.15.0
+> > git tag -a v2.17.0 -m "Ported StatusLine"</br>
+> > git push origin v2.17.0
 > 
 
 IMPORTANT: 
@@ -28,6 +28,15 @@ Correct URL for github, (for source compare)</br>
 Adding mock test cases </br>
 Adding unit test cases </br>
 Updated README.md accordinging</br>
+
+#### 2.17.0 (2021-11-19
+## Added  
+- Ported StatusLine
+
+#### 2.16.0 (2021-11-19
+## Added  
+- Documented Vendor/Server.cpp
+* [RSi-12]
 
 #### 2.15.0 (2021-11-18)
 ## Added  
@@ -178,7 +187,9 @@ Updated README.md accordinging</br>
 ### Removed
 - Conan package manager deprecated
 
-[Unreleased]: https://github.com/perriera/rsi/compare/v2.15.0...HEAD
+[Unreleased]: https://github.com/perriera/rsi/compare/v2.17.0...HEAD
+[2.17.0]: https://github.com/perriera/rsi/compare/v2.16.0...v2.17.0
+[2.16.0]: https://github.com/perriera/rsi/compare/v2.15.0...v2.16.0
 [2.15.0]: https://github.com/perriera/rsi/compare/v2.14.0...v2.15.0
 [2.14.0]: https://github.com/perriera/rsi/compare/v2.13.0...v2.14.0
 [2.13.1]: https://github.com/perriera/rsi/compare/v2.13.0...v2.13.1